requirements.txt
pandas
numpy
scikit-learn  == 1.6.1
xgboost       
scipy         
matplotlib
seaborn
imbalanced-learn
mlflow
fastapi
uvicorn[standard] 
pydantic          
psycopg2-binary
gradio
requests     
funpymodeling 
evidently
streamlit
